ref: Move OID mapper inside test namespace
