+++ tests/integration/Utils/DER/DerPublicKeySerializer.php
@@ -6,11 +6,11 @@ namespace Mdanter\Ecc\Integration\Utils\DER;
 use Mdanter\Ecc\Crypto\Key\PublicKey;
 use Mdanter\Ecc\Crypto\Key\PublicKeyInterface;
 use Mdanter\Ecc\Curves\NamedCurveFp;
+use Mdanter\Ecc\Integration\Utils\OID\CurveOidMapper;
 use Mdanter\Ecc\Math\GmpMathInterface;
 use Mdanter\Ecc\Math\MathAdapterFactory;
 use Mdanter\Ecc\Serializer\Point\ChainedPointSerializer;
 use Mdanter\Ecc\Serializer\Point\PointSerializerInterface;
-use Mdanter\Ecc\Serializer\Util\CurveOidMapper;
 use Sop\ASN1\Type\Constructed\Sequence;
 use Sop\ASN1\Type\Primitive\BitString;
 use Sop\ASN1\Type\Primitive\ObjectIdentifier;
